# project2-ecommerce/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "E-Commerce API",
    "description": "API documentation for Users, Products, and Orders endpoints with full CRUD operations",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "List of users",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/User" } }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User object to create",
            "required": true,
            "schema": { "$ref": "#/definitions/User" }
          }
        ],
        "responses": { "201": { "description": "User created" } }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get a single user by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Single user", "schema": { "$ref": "#/definitions/User" } },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "summary": "Update a user by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "user",
            "in": "body",
            "description": "Updated user data",
            "required": true,
            "schema": { "$ref": "#/definitions/User" }
          }
        ],
        "responses": {
          "200": { "description": "User updated" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "summary": "Delete a user by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "User deleted" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "List of products",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/Product" } }
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Product object to create",
            "required": true,
            "schema": { "$ref": "#/definitions/Product" }
          }
        ],
        "responses": { "201": { "description": "Product created" } }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get a single product by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Single product", "schema": { "$ref": "#/definitions/Product" } },
          "404": { "description": "Product not found" }
        }
      },
      "put": {
        "summary": "Update a product by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "product",
            "in": "body",
            "description": "Updated product data",
            "required": true,
            "schema": { "$ref": "#/definitions/Product" }
          }
        ],
        "responses": {
          "200": { "description": "Product updated" },
          "404": { "description": "Product not found" }
        }
      },
      "delete": {
        "summary": "Delete a product by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Product deleted" },
          "404": { "description": "Product not found" }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Get all orders",
        "responses": {
          "200": {
            "description": "List of orders",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/Order" } }
          }
        }
      },
      "post": {
        "summary": "Create a new order",
        "parameters": [
          {
            "name": "order",
            "in": "body",
            "description": "Order object to create",
            "required": true,
            "schema": { "$ref": "#/definitions/Order" }
          }
        ],
        "responses": { "201": { "description": "Order created" } }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get a single order by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Single order", "schema": { "$ref": "#/definitions/Order" } },
          "404": { "description": "Order not found" }
        }
      },
      "put": {
        "summary": "Update an order by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "order",
            "in": "body",
            "description": "Updated order data",
            "required": true,
            "schema": { "$ref": "#/definitions/Order" }
          }
        ],
        "responses": {
          "200": { "description": "Order updated" },
          "404": { "description": "Order not found" }
        }
      },
      "delete": {
        "summary": "Delete an order by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Order deleted" },
          "404": { "description": "Order not found" }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "firstName": { "type": "string" },
        "lastName": { "type": "string" },
        "email": { "type": "string" },
        "password": { "type": "string" },
        "role": { "type": "string" },
        "address": { "type": "string" },
        "phone": { "type": "string" },
        "createdAt": { "type": "string", "format": "date-time" }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "description": { "type": "string" },
        "price": { "type": "number" },
        "category": { "type": "string" },
        "stock": { "type": "number" },
        "imageUrl": { "type": "string" },
        "sku": { "type": "string" },
        "createdAt": { "type": "string", "format": "date-time" }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "userId": { "type": "string" },
        "products": {
          "type": "array",
          "items": { "$ref": "#/definitions/Product" }
        },
        "totalPrice": { "type": "number" },
        "status": { "type": "string" },
        "shippingAddress": { "type": "string" },
        "paymentMethod": { "type": "string" },
        "createdAt": { "type": "string", "format": "date-time" },
        "updatedAt": { "type": "string", "format": "date-time" }
      }
    }
  }
}